# Avoid allocations and a reflect.Call in conversion