--- a/pkg/conversion/deep_copy_generated.go
+++ b/pkg/conversion/deep_copy_generated.go
@@ -67,6 +67,19 @@ func DeepCopy_conversion_Converter(in Converter, out *Converter, c *Cloner) erro
 	if err := DeepCopy_conversion_ConversionFuncs(in.generatedConversionFuncs, &out.generatedConversionFuncs, c); err != nil {
 		return err
 	}
+	if in.genericConversions != nil {
+		in, out := in.genericConversions, &out.genericConversions
+		*out = make([]GenericConversionFunc, len(in))
+		for i := range in {
+			if newVal, err := c.DeepCopy(in[i]); err != nil {
+				return err
+			} else {
+				(*out)[i] = newVal.(GenericConversionFunc)
+			}
+		}
+	} else {
+		out.genericConversions = nil
+	}
 	if in.ignoredConversions != nil {
 		in, out := in.ignoredConversions, &out.ignoredConversions
 		*out = make(map[typePair]struct{})