@wojtek-t 2016-04-25
I'm fine with this PR.
I didn't have time to go some more sophisticated experiments, but my feeling is that if we also add Node and Event to this fast path, this can give us a looot.

Please remove "WIP" from the title, and self-apply lgtm label.