@smarterclayton 2016-04-25
I'm going to fiddle it up a bit to make it an array of matches (so we can
have each group register one) and then will set it up as you say.

On Mon, Apr 25, 2016 at 2:26 AM, Wojciech Tyczynski <
notifications@github.com> wrote:

> I'm fine with this PR.
> I didn't have time to go some more sophisticated experiments, but my
> feeling is that if we also add Node and Event to this fast path, this can
> give us a looot.
>
> Please remove "WIP" from the title, and self-apply lgtm label.
>
> —
> You are receiving this because you authored the thread.
> Reply to this email directly or view it on GitHub
> <https://github.com/kubernetes/kubernetes/pull/24727#issuecomment-214158468>
>