@smarterclayton 2016-04-25
Renamed this to be more consistent with what it is doing, added our N most converted types, and supported iteration.