@smarterclayton 2016-04-26
As per @wojtek-t applying LGTM