@wojtek-t 2016-04-26
One minor comment, if you can apply :)