@smarterclayton 2016-04-26
@k8s-bot test this issue #24211 for "scale from 2 pods to 1 pod using HPA version v1"