@smarterclayton 2016-04-26
@k8s-bot test this issue #24211 "Horizontal pod autoscaling"